requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.0,<5.0
django-environ>=0.11,<1.0
requests>=2.31,<3.0
gunicorn>=22.0,<23.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
openai>=1.0,<2.0
